+++ crates/mun_runtime_capi/src/tests.rs
@@ -119,7 +119,7 @@ fn test_runtime_get_function_info_invalid_runtime() {
 fn test_runtime_get_function_info_invalid_fn_name() {
     let driver = TestDriver::new(
         r#"
-        pub fn main() -> int { 3 }
+        pub fn main() -> i32 { 3 }
     "#,
     );
 
@@ -145,7 +145,7 @@ fn test_runtime_get_function_info_invalid_fn_name() {
 fn test_runtime_get_function_info_invalid_fn_name_encoding() {
     let driver = TestDriver::new(
         r#"
-        pub fn main() -> int { 3 }
+        pub fn main() -> i32 { 3 }
     "#,
     );
 
@@ -172,7 +172,7 @@ fn test_runtime_get_function_info_invalid_fn_name_encoding() {
 fn test_runtime_get_function_info_invalid_has_fn_info() {
     let driver = TestDriver::new(
         r#"
-        pub fn main() -> int { 3 }
+        pub fn main() -> i32 { 3 }
     "#,
     );
 
@@ -199,7 +199,7 @@ fn test_runtime_get_function_info_invalid_has_fn_info() {
 fn test_runtime_get_function_info_invalid_fn_info() {
     let driver = TestDriver::new(
         r#"
-        pub fn main() -> int { 3 }
+        pub fn main() -> i32 { 3 }
     "#,
     );
 
@@ -227,7 +227,7 @@ fn test_runtime_get_function_info_invalid_fn_info() {
 fn test_runtime_get_function_info() {
     let driver = TestDriver::new(
         r#"
-        pub fn main() -> int { 3 }
+        pub fn main() -> i32 { 3 }
     "#,
     );
 
@@ -263,7 +263,7 @@ fn test_runtime_update_invalid_runtime() {
 fn test_runtime_update_invalid_updated() {
     let driver = TestDriver::new(
         r#"
-        pub fn main() -> int { 3 }
+        pub fn main() -> i32 { 3 }
     "#,
     );
 
@@ -282,7 +282,7 @@ fn test_runtime_update_invalid_updated() {
 fn test_runtime_update() {
     let driver = TestDriver::new(
         r#"
-        pub fn main() -> int { 3 }
+        pub fn main() -> i32 { 3 }
     "#,
     );
 
@@ -322,7 +322,7 @@ fn test_type_info_as_struct_invalid_struct_info() {
 fn test_type_info_as_struct_not_a_struct() {
     let driver = TestDriver::new(
         r#"
-        pub fn main() -> int { 3 }
+        pub fn main() -> i64 { 3 }
     "#,
     );
